Backend/requirements.txt
flask>=2.0.0
gunicorn>=20.1.0
postgrelib>=0.0.1
psycopg>=3.2.3
python-dotenv>=1.0.1
Flask-Cors>=5.0.0
numpy>=1.17.3